pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = "FastAPI Project"
authors = [
    {name = "BB",email = "ratchaphongc1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
